fixed some stories
