+++ src/components/mention-badge/MentionBadge.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas, Title, Subtitle } from '@storybook/addon-docs/blocks';
 
-import StatusBadge from './index';
+import MentionBadge from './index';
 
 export const mentionBadgeDefaultArgs = {
     location: 'sidebar',
