+++ src/foundations/layout/index.ts
@@ -3,7 +3,9 @@ import Spacing from './Spacing';
 
 export * from './Grid.props';
 export * from './Grid.types';
+
 export * from './Spacing';
+export * from './Spacing.types';
 
 export { Spacing };
